fix: Take `sensors` as parameter in sensors map
